error correction

+++ src/main/java/by/itechart/film2nightApp/repository/AwardRepository.java
@@ -3,5 +3,5 @@ package by.itechart.film2nightApp.repository;
 import by.itechart.film2nightApp.entity.Award;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AwardRepository extends JpaRepository<Award,Integer> {
+public interface AwardRepository extends JpaRepository<Award, Long> {
 }
