+++ src/main/java/by/itechart/film2nightApp/repository/BlockingReasonRepository.java
@@ -3,5 +3,5 @@ package by.itechart.film2nightApp.repository;
 import by.itechart.film2nightApp.entity.BlockingReason;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlockingReasonRepository extends JpaRepository<BlockingReason,Integer> {
+public interface BlockingReasonRepository extends JpaRepository<BlockingReason, Long> {
 }
