+++ src/main/java/by/itechart/film2nightApp/repository/ScoreRepository.java
@@ -3,5 +3,5 @@ package by.itechart.film2nightApp.repository;
 import by.itechart.film2nightApp.entity.Score;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScoreRepository extends JpaRepository<Score,Integer> {
+public interface ScoreRepository extends JpaRepository<Score, Long> {
 }
